Time: 0 ms (100%), Space: 9 MB (30.63%) - LeetHub

+++ 0268-missing-number/0268-missing-number.c
@@ -1,5 +1,5 @@
 int missingNumber(int* t, int n) { 
-    unsigned long int v = ((n+1)*n)/2;
+    unsigned short int v = ((n+1)*n)/2;
     for(unsigned short int i=0; i<n; i++){
         v -= t[i];
     }
